move types to one file
