--- src/typings/helpers.d.ts
@@ -1,17 +1,0 @@
-import type {
-    EmbedBuilder,
-    Guild,
-    ColorResolvable,
-    Attachment,
-    AttachmentBuilder 
-} from 'discord.js';
-
-interface BotLogOptions {
-    guild: Guild,
-    data: { 
-        title: string,
-        color?: ColorResolvable, 
-        embed?: EmbedBuilder,
-        attachments?: (Attachment | AttachmentBuilder)[]
-    }
-}
