+++ src/typings/index.d.ts
@@ -1,8 +1,13 @@
 import type {
     AutocompleteInteraction,
+    Attachment,
+    AttachmentBuilder ,
     ChatInputCommandInteraction,
     Collection,
+    ColorResolvable,
+    EmbedBuilder,
     Events,
+    Guild,
     SlashCommandBuilder
 } from 'discord.js';
 import type {
@@ -10,6 +15,22 @@ import type {
     Kysely
 } from 'kysely';
 import type { DB } from './database.js';
+
+// #region Modules
+
+declare module 'discord.js' {
+    export interface Client {
+        commands: Collection<string, Command>,
+        db: Kysely<DB>,
+        util: Collection<string, Utility>
+        getConfig(option: string, guild?: string): string | undefined,
+        refreshConfig(): Promise<void>
+    }
+}
+
+// #endregion Modules
+
+// #region Interfaces
 
 export interface Command {
     data: SlashCommandBuilder,
@@ -24,12 +45,14 @@ export interface Utility {
     execute: (...args: any) => void
 }
 
-declare module 'discord.js' {
-    export interface Client {
-        commands: Collection<string, Command>,
-        db: Kysely<DB>,
-        util: Collection<string, Utility>
-        getConfig(option: string, guild?: string): string | undefined,
-        refreshConfig(): Promise<void>
+export interface BotLogOptions {
+    guild: Guild,
+    data: { 
+        title: string,
+        color?: ColorResolvable, 
+        embed?: EmbedBuilder,
+        attachments?: (Attachment | AttachmentBuilder)[]
     }
 }
+
+// #endregion Interfaces
